Modifications & Fixes

+++ src/public/js/main.js
@@ -1,39 +1,100 @@
 const usersBtnUpdate = document.querySelectorAll(".users-btn-update");
 const usersBtnDelete = document.querySelectorAll(".users-btn-delete");
-
 const crtUserImgUpload = document.getElementById("crt-user-img-upload");
 const crtUserImg = document.getElementById("crt-user-img");
+const crtUserForm = document.querySelector(".crt-user-form");
+const crtUserFormSubmit = document.querySelector(".crt-user-submit button");
+const updateUserFormSubmit = document.querySelector(".update-user-submit button");
 const userImgMessage = document.getElementById("crt-user-img-msg");
 const mobileUserCards = document.querySelectorAll(".mb-user-card");
+const btnDeleteMobileCard = document.querySelectorAll(".mb-user-card-btn-delete");
+const crtUserEmail = document.querySelector("#crt-user-email");
+const crtUserFirstname = document.querySelector("#crt-user-firstname");
+const crtUserLastname = document.querySelector("#crt-user-lastname");
+const userID = document.querySelector("#user-id");
 
-//User overview - update user data
+/* Create new user - User Form */
+if (crtUserFormSubmit)
+  crtUserFormSubmit.addEventListener("click", () => {
+    crtUserForm.method = "post";
+    crtUserForm.action = "/users/create";
+    crtUserForm.submit();
+    crtUserForm.reset();
+  });
+
+/* Update existing user - User Form */
+if (updateUserFormSubmit)
+  updateUserFormSubmit.addEventListener("click", () => {
+    crtUserForm.method = "post";
+    crtUserForm.action = `/users/update/${userID.textContent}`;
+    crtUserForm.submit();
+    crtUserForm.reset();
+  });
+
+/* User Form - Check for valid email - Input EventListener */
+if (crtUserEmail)
+  crtUserEmail.addEventListener("focusout", (e) => {
+    if (!checkEmail(e.target.value)) {
+      crtUserEmail.classList.add("error-input");
+      return;
+    }
+    crtUserEmail.classList.remove("error-input");
+  });
+
+/* Check for valid email */
+const checkEmail = (value) => {
+  const regex = new RegExp("[a-z0-9._%+-]+@[a-z0-9.-]+.[a-z]{2,4}$");
+  if (regex.test(value)) {
+    return true;
+  }
+  return false;
+};
+
+/* User Form - Check input values */
+if (crtUserFirstname)
+  crtUserFirstname.addEventListener("focusout", (e) => {
+    if (e.target.value === "") {
+      crtUserFirstname.classList.add("error-input");
+      return;
+    }
+    crtUserFirstname.classList.remove("error-input");
+  });
+
+if (crtUserLastname)
+  crtUserLastname.addEventListener("focusout", (e) => {
+    if (e.target.value === "") {
+      crtUserLastname.classList.add("error-input");
+      return;
+    }
+    crtUserLastname.classList.remove("error-input");
+  });
+
+/* Handling update user */
 usersBtnUpdate.forEach((btn) =>
   btn.addEventListener("click", async (e) => {
     const record = btn.parentElement.parentElement;
-    const id = btn.parentElement.parentElement.querySelector(".users-th-id").innerHTML;
-    const email = btn.parentElement.parentElement.querySelector(".users-td-email").innerHTML;
-    const firstname = btn.parentElement.parentElement.querySelector(".users-td-firstname").innerHTML;
-    const lastname = btn.parentElement.parentElement.querySelector(".users-td-lastname").innerHTML;
-    const updated = btn.parentElement.parentElement.querySelector(".users-td-updated");
-    //Ajax request
+    const id = record.querySelector(".users-th-id").textContent;
+    const email = record.querySelector(".users-td-email").textContent;
+    const firstname = record.querySelector(".users-td-firstname").textContent;
+    const lastname = record.querySelector(".users-td-lastname").textContent;
+    const updated = record.querySelector(".users-td-updated");
     await updateUser({ id, email, firstname, lastname })
       .then((response) => {
-        //Check response code
         if (response.ok) {
           //Success animation
-          record.classList.add("record-success");
+          record.classList.add("users-record-success");
           record.addEventListener("animationend", () => {
-            record.classList.remove("record-success");
+            record.classList.remove("users-record-success");
           });
-          //Extract data
           response.json().then((data) => {
-            //Update Form timestamp
-            updated.innerHTML = data[0].updated;
+            //Set updated timestamp in user record
+            updated.innerHTML = data.formattedUpdated;
+            //Reload to update mobile view
+            location.reload();
           });
         }
       })
       .catch((error) => {
-        console.log(error);
         //Error animation
         record.classList.add("users-record-error");
         record.addEventListener("animationend", () => {
@@ -56,18 +117,34 @@ async function updateUser(user) {
   }
 }
 
-//Delete user data - Ajax
+/* Handling delete user - Desktop View */
 usersBtnDelete.forEach((btn) =>
   btn.addEventListener("click", async (e) => {
-    //Get user
     const record = btn.parentElement.parentElement;
-    const id = btn.parentElement.parentElement.querySelector(".users-th-id").innerHTML;
-    //Ajax request
+    const id = record.querySelector(".users-th-id").innerHTML;
     await deleteUser(id)
       .then((response) => {
-        //Check response code - remove user record
         if (response.ok) {
           record.remove();
+          location.reload();
+        }
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  })
+);
+
+/* Handling delete user - Mobile View */
+btnDeleteMobileCard.forEach((btn) =>
+  btn.addEventListener("click", async (e) => {
+    const userCard = btn.parentElement.parentElement.parentElement;
+    const id = userCard.querySelector("#mb-user-card-id").innerHTML;
+    await deleteUser(id)
+      .then((response) => {
+        if (response.ok) {
+          userCard.remove();
+          location.reload();
         }
       })
       .catch((error) => {
@@ -77,37 +154,38 @@ usersBtnDelete.forEach((btn) =>
 );
 
 async function deleteUser(id) {
-  try {
-    const response = await fetch(`/users/${id}`, {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-    });
-    return response;
-  } catch (error) {
-    return error;
+  // Delete confirmation
+  if (window.confirm("Do you really want to delete this user?")) {
+    try {
+      const response = await fetch(`/users/${id}`, {
+        method: "DELETE",
+        headers: { "Content-Type": "application/json" },
+      });
+      return response;
+    } catch (error) {
+      return error;
+    }
   }
 }
 
-//User image select
-crtUserImgUpload.addEventListener("change", userImgFileHandler);
+/* Handling user image select */
+if (crtUserImgUpload) crtUserImgUpload.addEventListener("change", userImgFileHandler);
 
 function userImgFileHandler(event) {
   //Check if a file has been selected
   if (event.target.files.length === 0) return;
   //Get selected image. Check data type
   const userImg = event.target.files[0];
-
   if (!checkFileType(event.target.files[0].type)) {
     userImgMessage.textContent = "Please select an image file.";
     return;
   }
-
-  //Check size
+  //Check file size
   if (userImg.size > 200000) {
-    userImgMessage.textContent = "Max supported image size: < 200Kb.";
+    window.alert("Max supported image size: < 200Kb.");
     return;
   }
-  //Select & remove default avatar
+  //Remove default avatar if an image has been selected
   const avatar = document.querySelector(".crt-user-avatar");
   if (avatar) avatar.remove();
   //Create image
@@ -116,7 +194,7 @@ function userImgFileHandler(event) {
   crtUserImg.appendChild(image);
 }
 
-//Check file type for user profile image
+/* File handling for user image select */
 function checkFileType(type) {
   const fileTypes = [
     "image/apng",
@@ -134,7 +212,7 @@ function checkFileType(type) {
   return fileTypes.includes(type);
 }
 
-//Mobile View User Edit
+/* Control User Edit - Mobile View */
 mobileUserCards.forEach((element) => {
   element.addEventListener("click", (ev) => {
     element.children[0].classList.toggle("mb-user-card-control-expand");
